refactor: mark CalendarLib as class library
Since the CalendarLib project will be used as a class library for
frontends to use, no main method is required, delete that.

Also mark it as a class library in the .csproj file and the .sln file
for the whole solution, so the build system properly understands it, and
builds it as a library.

Author: Thorben von Rekowski <thorben.von-rekowski@gwdg.de>